+++ app/Models/User.php
@@ -2,47 +2,44 @@
 
 namespace App\Models;
 
-// use Illuminate\Contracts\Auth\MustVerifyEmail;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 
 class User extends Authenticatable
 {
-    /** @use HasFactory<\Database\Factories\UserFactory> */
-    use HasFactory, Notifiable;
+    use Notifiable;
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var list<string>
-     */
+    protected $table = 'users';
+    protected $primaryKey = 'user_id';
+
     protected $fillable = [
-        'name',
+        'user_name', 
         'email',
+        'phone',
         'password',
+        'country',
+        'role_id',
     ];
 
-    /**
-     * The attributes that should be hidden for serialization.
-     *
-     * @var list<string>
-     */
-    protected $hidden = [
-        'password',
-        'remember_token',
-    ];
 
-    /**
-     * Get the attributes that should be cast.
-     *
-     * @return array<string, string>
-     */
-    protected function casts(): array
+  public function role(){
+      return $this->belongsTo(Role::class, 'role_id');}
+
+
+  public function jobposts(){
+      return $this->hasMany(JobPost::class, 'user_id');}
+
+
+  public function notifications(){
+      return $this->hasMany(Notification::class, 'user_id');}
+
+
+  public function reports(){
+      return $this->hasMany(Report::class, 'user_id');
+    }
+
+    public function profile()
     {
-        return [
-            'email_verified_at' => 'datetime',
-            'password' => 'hashed',
-        ];
+    return $this->hasOne(Profile::class, 'user_id');
     }
 }
